Replace AddQuestion screen with dialog
