+++ app/src/main/java/com/cihat/egitim/lottieanimation/ui/screens/QuizListScreen.kt
@@ -126,8 +126,8 @@ fun QuizListScreen(
     onMoveQuiz: (Int, Int) -> Unit,
     onCreate: (String, Int, Int?) -> Unit,
     onCreateWithQuestion: (String, Int, Int?, String, String, String, String) -> Unit,
-    onQuickAdd: (Int, String, String, String, String) -> Unit,
-    onAddQuestion: (Int) -> Unit,
+    onAddQuestion: (String, String, String, String, Int) -> Unit,
+    onSetCurrentQuiz: (Int) -> Unit,
     onFolders: () -> Unit,
     onBack: () -> Unit,
     onTab: (BottomTab) -> Unit
@@ -152,6 +152,7 @@ fun QuizListScreen(
         val itemHeightPx = with(LocalDensity.current) { 72.dp.toPx() }
         var startDialogFor by remember { mutableStateOf<Int?>(null) }
         var emptyAlertFor by remember { mutableStateOf<Int?>(null) }
+        var addDialogFor by remember { mutableStateOf<Int?>(null) }
 
         // State for draggable FAB
         val density = LocalDensity.current
@@ -209,10 +210,6 @@ fun QuizListScreen(
                             var expanded by remember(quiz.id) { mutableStateOf(false) }
                             var showRename by remember(quiz.id) { mutableStateOf(false) }
                             var showDelete by remember(quiz.id) { mutableStateOf(false) }
-                            var showQuickAdd by remember(quiz.id) { mutableStateOf(false) }
-                            var quickQuestion by remember(quiz.id) { mutableStateOf("") }
-                            var quickAnswer by remember(quiz.id) { mutableStateOf("") }
-                            var quickPath by remember(quiz.id) { mutableStateOf<List<Int>>(emptyList()) }
                             var newName by remember(quiz.id) { mutableStateOf(quiz.name) }
                             val scope = rememberCoroutineScope()
                             val actionWidth = 72.dp
@@ -326,7 +323,8 @@ fun QuizListScreen(
                                     IconButton(
                                         onClick = {
                                             scope.launch { swipeState.animateTo(0) }
-                                            showQuickAdd = true
+                                            onSetCurrentQuiz(quizIndex)
+                                            addDialogFor = quizIndex
                                         },
                                         enabled = swipeState.currentValue == 2,
                                         modifier = Modifier
@@ -474,95 +472,17 @@ fun QuizListScreen(
                                 )
                             }
 
-                            if (showQuickAdd) {
+                            if (addDialogFor == quizIndex) {
                                 val folderId = quiz.folderId
                                 val folderHeadings = folders.find { it.id == folderId }?.headings
                                     ?: headingsFromQuestions(quiz.boxes.flatten())
-                                AlertDialog(
-                                    onDismissRequest = { showQuickAdd = false },
-                                    confirmButton = {
-                                        TextButton(onClick = {
-                                            var list = folderHeadings
-                                            val names = mutableListOf<String>()
-                                            for (idx in quickPath) {
-                                                val h = list.getOrNull(idx) ?: break
-                                                names.add(h.name)
-                                                list = h.children
-                                            }
-                                            val topic = names.firstOrNull() ?: ""
-                                            val sub = names.drop(1).joinToString(" > ")
-                                            onQuickAdd(quizIndex, topic, sub, quickQuestion, quickAnswer)
-                                            showQuickAdd = false
-                                            quickQuestion = ""
-                                            quickAnswer = ""
-                                            quickPath = emptyList()
-                                        }) { Text("Soru Ekle") }
+                                AddQuestionDialog(
+                                    boxCount = quiz.boxes.size,
+                                    headings = folderHeadings,
+                                    onAdd = { q, a, topic, sub, box ->
+                                        onAddQuestion(q, a, topic, sub, box)
                                     },
-                                    dismissButton = {
-                                        TextButton(onClick = { showQuickAdd = false }) { Text("İptal") }
-                                    },
-                                    title = {
-                                        Text(folderId?.let { folders.find { f -> f.id == it }?.name }
-                                            ?: "")
-                                    },
-                                    text = {
-                                        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                                            var currentList = folderHeadings
-                                            for (level in 0..quickPath.size) {
-                                                val options = currentList
-                                                if (options.isEmpty()) break
-                                                var expanded by remember(level, quickPath) { mutableStateOf(false) }
-                                                val selectedIdx = quickPath.getOrNull(level)
-                                                ExposedDropdownMenuBox(
-                                                    expanded = expanded,
-                                                    onExpandedChange = { expanded = !expanded }
-                                                ) {
-                                                    OutlinedTextField(
-                                                        value = selectedIdx?.let { options[it].name }
-                                                            ?: "Seçiniz",
-                                                        onValueChange = {},
-                                                        readOnly = true,
-                                                        label = { Text("Başlık ${level + 1}") },
-                                                        trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded) },
-                                                        modifier = Modifier
-                                                            .menuAnchor()
-                                                            .fillMaxWidth()
-                                                    )
-                                                    ExposedDropdownMenu(
-                                                        expanded = expanded,
-                                                        onDismissRequest = { expanded = false }
-                                                    ) {
-                                                        options.forEachIndexed { index, h ->
-                                                            DropdownMenuItem(
-                                                                text = { Text(h.name) },
-                                                                onClick = {
-                                                                    quickPath = quickPath.take(level) + index
-                                                                    expanded = false
-                                                                }
-                                                            )
-                                                        }
-                                                    }
-                                                }
-                                                Spacer(modifier = Modifier.height(8.dp))
-                                                selectedIdx?.let { idx -> currentList = options[idx].children }
-                                                    ?: run { currentList = emptyList() }
-                                            }
-
-                                            OutlinedTextField(
-                                                value = quickQuestion,
-                                                onValueChange = { quickQuestion = it },
-                                                label = { Text("Soru") },
-                                                modifier = Modifier.fillMaxWidth()
-                                            )
-                                            Spacer(modifier = Modifier.height(8.dp))
-                                            OutlinedTextField(
-                                                value = quickAnswer,
-                                                onValueChange = { quickAnswer = it },
-                                                label = { Text("Cevap") },
-                                                modifier = Modifier.fillMaxWidth()
-                                            )
-                                        }
-                                    }
+                                    onDismiss = { addDialogFor = null }
                                 )
                             }
 
@@ -706,7 +626,8 @@ fun QuizListScreen(
                 confirmText = "Soru Ekle",
                 onConfirm = {
                     emptyAlertFor = null
-                    onAddQuestion(idx)
+                    onSetCurrentQuiz(idx)
+                    addDialogFor = idx
                 }
             )
         }
